+++ packages/harmony/src/components/button/FollowButton/FollowButton.mdx
@@ -1,0 +1,43 @@
+import {
+  Canvas,
+  Meta,
+  Title,
+  Description,
+  Primary,
+  Controls
+} from '@storybook/blocks'
+
+import {
+  Heading,
+  ComponentRules,
+  RelatedComponents
+} from 'storybook/components'
+
+import * as FollowButtonStories from './FollowButton.stories'
+
+<Meta of={FollowButtonStories} />
+
+<Title />
+
+- [Overview](#overview)
+- [Props](#props)
+- [Configuration Options](#configuration-options)
+
+<Heading>
+  ## Overview
+  <Description />
+</Heading>
+
+## Props
+
+<Controls />
+
+## Configuration Options
+
+### Variants
+
+<Canvas of={FollowButtonStories.Variants} />
+
+### Sizes
+
+<Canvas of={FollowButtonStories.Sizes} />
