+++ packages/harmony/src/components/button/FollowButton/types.ts
@@ -1,0 +1,30 @@
+import type { HTMLProps } from 'react'
+
+export type FollowButtonProps = {
+  /**
+   * Swap between the default (squared) and pill (rounded) forms.
+   * @default default
+   */
+  variant?: 'default' | 'pill'
+
+  /**
+   * Swap between the default and small sizes.
+   * @default default
+   */
+  size?: 'default' | 'small'
+
+  /**
+   * The current state of the button
+   */
+  following?: boolean
+
+  /**
+   * Callback for when a follow is triggered.
+   */
+  onFollow?: () => void
+
+  /**
+   * Callback for when an unfollow is triggered.
+   */
+  onUnfollow?: () => void
+} & Omit<HTMLProps<HTMLInputElement>, 'size'>
